vuex objects must be plan object as vue data
