+++ tests/unit/model/user.model.spec.js
@@ -1,6 +1,4 @@
 import { expect } from 'chai'
-// import { shallowMount } from '@vue/tests-utils'
-// import HelloWorld from '@/components/HelloWorld.vue'
 import User from '@/model/User.model.js'
 
 describe('UserModel', () => {
@@ -18,20 +16,5 @@ describe('UserModel', () => {
       expect(user.created).to.equal(created)
       expect(user.server).to.equal(server)
     })
-  })
-  describe('Hub URL', () => {
-    const user = new User('', [], '', false, '')
-    expect(user.getHubUrl()).to.equal('/hub/home')
-    user.server = '/user/john.foo/'
-    expect(user.getHubUrl()).to.equal('/hub/home')
-    user.server = '/cylc/user/john.foo/'
-    expect(user.getHubUrl()).to.equal('/cylc/hub/home')
-    user.server = '/cylc/flow/workflow/engine/user/john.foo/'
-    expect(user.getHubUrl()).to.equal('/cylc/flow/workflow/engine/hub/home')
-    // who knows right? no idea if the configurable-http-proxy would still work with these URLs, but just in case...
-    user.server = '/user/user/john.foo/'
-    expect(user.getHubUrl()).to.equal('/user/hub/home')
-    user.server = '/cylc/flow/user/workflow/engine/user/john.foo/'
-    expect(user.getHubUrl()).to.equal('/cylc/flow/user/workflow/engine/hub/home')
   })
 })
